+++ spinta/formats/ascii/commands.py
@@ -2,6 +2,7 @@ import subprocess
 import sys
 
 from typing import Optional
+from typing import List
 
 from starlette.requests import Request
 from starlette.responses import StreamingResponse
@@ -25,10 +26,19 @@ def render(
     status_code: int = 200,
     headers: Optional[dict] = None,
 ):
-    return _render(fmt, params, data, status_code, headers)
+    return _render(context, model, fmt, action, params, data, status_code, headers)
 
 
-def _render(fmt: Ascii, params: UrlParams, data, status_code, headers):
+def _render(
+    context: Context,
+    model: Model,
+    fmt: Ascii,
+    action: Action,
+    params: UrlParams,
+    data,
+    status_code,
+    headers,
+):
     # Format params ar given in RQL query `?format(width(1),colwidth(1))`.
     width = params.formatparams.get('width')
     colwidth = params.formatparams.get('colwidth')
@@ -39,7 +49,7 @@ def _render(fmt: Ascii, params: UrlParams, data, status_code, headers):
         colwidth = 42
 
     return StreamingResponse(
-        aiter(fmt(data, width, colwidth)),
+        aiter(fmt(context, model, action, params, data, width, colwidth)),
         status_code=status_code,
         media_type=fmt.content_type,
         headers=headers,