+++ spinta/formats/ascii/components.py
@@ -1,10 +1,12 @@
 import operator
-
 import itertools
 
+from spinta.components import Context, Action, UrlParams, Model
+from spinta.manifests.components import Manifest
 from spinta.formats.components import Format
 from spinta.formats.ascii.helpers import draw
 from spinta.utils.nestedstruct import flatten
+from spinta.formats.helpers import get_model_tabular_header
 
 
 class Ascii(Format):
@@ -17,7 +19,18 @@ class Ascii(Format):
         'colwidth': {'type': 'integer'},
     }
 
-    def __call__(self, data, width=None, colwidth=42):
+    def __call__(
+        self,
+        context: Context,
+        model: Model,
+        action: Action,
+        params: UrlParams,
+        data,
+        width=None,
+        colwidth=42,
+    ):
+        manifest: Manifest = context.get('store').manifest
+
         data = iter(data)
         peek = next(data, None)
 
@@ -34,15 +47,10 @@ class Ascii(Format):
         for name, group in groups:
             if name:
                 yield f'\n\nTable: {name}\n'
+                model = manifest.models[name]
 
             rows = flatten(group)
-            peek = next(rows, None)
-            rows = itertools.chain([peek], rows)
-
-            if name:
-                cols = [k for k in peek.keys() if k != '_type']
-            else:
-                cols = list(peek.keys())
+            cols = get_model_tabular_header(context, model, action, params)
 
             if colwidth:
                 width = len(cols) * colwidth